requirements.txt
# Core dependencies
psycopg2-binary==2.9.9
openai>=1.0.0
numpy>=1.24.0
tqdm==4.66.5
requests==2.31.0
pgvector>=0.2.0
